✨ feat: (FE) - selector가 webrtc와 user 정보를 통합하여 표시하도록 변경
